Uploads Manager: Moved the Uploads Manager to the core directory.
